Update .bashrc
Added better CDPATH, a couple aliases, cleaned up the file. Trying to add a bit of normalization IE: 4 hashtags on main comments and a common schema for the information displayed at the top of the configuration file. Whether this is important to this project I'm not yet sure, but I like the more organized feel.